+++ examples/cache.ts
@@ -1,4 +1,4 @@
-import { config } from "snackables";
+import { config, getCache } from "snackables";
 
 console.log(`\x1b[32mCache is working if warning is thrown...\x1b[0m`);
 
@@ -6,4 +6,10 @@ const { cachedENVFiles } = config({ path: "base" });
 
 console.log(
   `\x1b[32mcachedENVFiles: ${JSON.stringify(cachedENVFiles, null, 2)}\x1b[0m`
+);
+
+console.log(
+  `\x1b[32mcachedENVFiles === getCache(): ${
+    cachedENVFiles === getCache()
+  }\x1b[0m`
 );