+++ examples/customdir.ts
@@ -1,8 +1,7 @@
 import { config } from "snackables";
 
-const { parsed, extracted } = config();
+const { extracted } = config({ dir: "tests" });
 
-console.log(`\x1b[32mparsed ENVs: ${JSON.stringify(parsed, null, 2)}\x1b[0m`);
 console.log(
   `\x1b[32mextracted ENVs: ${JSON.stringify(extracted, null, 2)}\x1b[0m`
 );