+++ index.d.ts
@@ -13,11 +13,13 @@ export type Encoding =
   | "binary"
   | "hex";
 
-interface ParsedOutput {
+interface ParsedENVs {
   [name: string]: string;
 }
 
-type CachedENVFiles = Array<{
+export type ProcessEnv = { [key: string]: string };
+
+export type ENVFiles = Array<{
   path: string;
   contents: any;
 }>;
@@ -25,31 +27,37 @@ type CachedENVFiles = Array<{
 /**
  * Parses a string or buffer into an object of ENVs.
  *
- * @param src - string or Buffer to be parsed
+ * @param src - string or Buffer or an array of precached files to be parsed
  * @returns a single object with parsed ENVs as { key: value } pairs
  */
-
-export function parse(src: string | Buffer): ParsedOutput;
+export function parse(src: string | Buffer | ENVFiles): ParsedENVs;
 
 export interface ConfigOptions {
+  dir?: string; // root directory
   path?: string | string[]; // path to .env file
   encoding?: Encoding; // encoding of .env file
   debug?: string | boolean; // turn on logging for debugging purposes
 }
 
 export interface ConfigOutput {
-  parsed: ParsedOutput; // parsed ENVs as key value pairs
-  cachedENVFiles: CachedENVFiles; // cached ENVs as key value pairs
+  parsed: ParsedENVs; // process.env ENVs as key value pairs
+  extracted: ParsedENVs; // extracted ENVs as key value pairs
+  cachedENVFiles: ENVFiles; // cached ENVs as key value pairs
 }
 
-function getCache(): CachedENVFiles;
+/**
+ * Returns internal cached ENVs files
+ *
+ * @returns an array of cached ENVs as { path: path/to/file, contents: file contents as string} pairs
+ */
+export function getCache(): ENVFiles;
 
 /**
  * Extracts and interpolates one or multiple `.env` files into an object and assigns them to {@link https://nodejs.org/api/process.html#process_process_env | `process.env`}.
  * Example: 'KEY=value' becomes { KEY: 'value' }
  *
- * @param options - accepts: { path: string | string[], debug: boolean, encoding: | "ascii" | "utf8" | "utf-8" | "utf16le" | "ucs2" | "ucs-2" | "base64" | "latin1" | "binary"| "hex" }
- * @returns a single parsed object with parsed ENVs as { key: value } pairs and an array of cached ENVs as { key: value} pairs
+ * @param options - accepts: { dir: string, path: string | string[], debug: boolean, encoding: | "ascii" | "utf8" | "utf-8" | "utf16le" | "ucs2" | "ucs-2" | "base64" | "latin1" | "binary"| "hex" }
+ * @returns a single parsed object with parsed ENVs as { key: value } pairs, a single extracted object with extracted ENVS as { key: value } pairs, and an array of cached ENVs as { path: string, contents: string} pairs
  */
 export function config(options?: ConfigOptions): ConfigOutput;